+++ appinfo/application.php
@@ -24,6 +24,7 @@ use OCP\AppFramework\Http\ContentSecurityPolicy;
 use OCP\DirectEditing\RegisterDirectEditorEvent;
 use OCP\Files\IMimeTypeDetector;
 use OCP\Util;
+use OCP\IPreview;
 
 use OCA\Viewer\Event\LoadViewer;
 
@@ -34,6 +35,7 @@ use OCA\Onlyoffice\Controller\SettingsController;
 use OCA\Onlyoffice\Crypt;
 use OCA\Onlyoffice\DirectEditor;
 use OCA\Onlyoffice\Hooks;
+use OCA\Onlyoffice\Preview;
 
 class Application extends App {
 
@@ -124,6 +126,10 @@ class Application extends App {
         $detector->registerType("ots", "application/vnd.oasis.opendocument.spreadsheet-template");
         $detector->registerType("otp", "application/vnd.oasis.opendocument.presentation-template");
 
+        $previewManager = $container->query(IPreview::class);
+        $previewManager->registerProvider(Preview::getMimeTypeRegex(), function() use ($container) {
+            return $container->query(Preview::class);
+        });
 
         $container->registerService("L10N", function ($c) {
             return $c->query("ServerContainer")->getL10N($c->query("AppName"));