+++ controller/settingscontroller.php
@@ -113,6 +113,7 @@ class SettingsController extends Controller {
             "currentServer" => $this->urlGenerator->getAbsoluteURL("/"),
             "formats" => $this->config->FormatsSetting(),
             "sameTab" => $this->config->GetSameTab(),
+            "preview" => $this->config->GetPreview(),
             "limitGroups" => $this->config->GetLimitGroups(),
             "chat" => $this->config->GetCustomizationChat(),
             "compactHeader" => $this->config->GetCustomizationCompactHeader(),
@@ -163,7 +164,8 @@ class SettingsController extends Controller {
         if (empty($error)) {
             $documentserver = $this->config->GetDocumentServerUrl();
             if (!empty($documentserver)) {
-                list ($error, $version) = $this->checkDocServiceUrl();
+                $documentService = new DocumentService($this->trans, $this->config);
+                list ($error, $version) = $documentService->checkDocServiceUrl($this->urlGenerator, $this->crypt);
                 $this->config->SetSettingsError($error);
             }
         }
@@ -185,6 +187,7 @@ class SettingsController extends Controller {
      * @param array $defFormats - formats array with default action
      * @param array $editFormats - editable formats array
      * @param bool $sameTab - open in the same tab
+     * @param bool $preview - generate preview files
      * @param array $limitGroups - list of groups
      * @param bool $chat - display chat
      * @param bool $compactHeader - display compact header
@@ -199,6 +202,7 @@ class SettingsController extends Controller {
     public function SaveCommon($defFormats,
                                     $editFormats,
                                     $sameTab,
+                                    $preview,
                                     $limitGroups,
                                     $chat,
                                     $compactHeader,
@@ -212,6 +216,7 @@ class SettingsController extends Controller {
         $this->config->SetDefaultFormats($defFormats);
         $this->config->SetEditableFormats($editFormats);
         $this->config->SetSameTab($sameTab);
+        $this->config->SetPreview($preview);
         $this->config->SetLimitGroups($limitGroups);
         $this->config->SetCustomizationChat($chat);
         $this->config->SetCustomizationCompactHeader($compactHeader);
@@ -261,88 +266,5 @@ class SettingsController extends Controller {
             "sameTab" => $this->config->GetSameTab()
         ];
         return $result;
-    }
-
-
-    /**
-     * Checking document service location
-     *
-     * @return array
-     */
-    private function checkDocServiceUrl() {
-        $version = null;
-
-        try {
-
-            if (preg_match("/^https:\/\//i", $this->urlGenerator->getAbsoluteURL("/"))
-                && preg_match("/^http:\/\//i", $this->config->GetDocumentServerUrl())) {
-                throw new \Exception($this->trans->t("Mixed Active Content is not allowed. HTTPS address for Document Server is required."));
-            }
-
-        } catch (\Exception $e) {
-            $this->logger->logException($e, ["message" => "Protocol on check error", "app" => $this->appName]);
-            return [$e->getMessage(), $version];
-        }
-
-        try {
-
-            $documentService = new DocumentService($this->trans, $this->config);
-
-            $healthcheckResponse = $documentService->HealthcheckRequest();
-            if (!$healthcheckResponse) {
-                throw new \Exception($this->trans->t("Bad healthcheck status"));
-            }
-
-        } catch (\Exception $e) {
-            $this->logger->logException($e, ["message" => "HealthcheckRequest on check error", "app" => $this->appName]);
-            return [$e->getMessage(), $version];
-        }
-
-        try {
-
-            $documentService = new DocumentService($this->trans, $this->config);
-
-            $commandResponse = $documentService->CommandRequest("version");
-
-            $this->logger->debug("CommandRequest on check: " . json_encode($commandResponse), ["app" => $this->appName]);
-
-            if (empty($commandResponse)) {
-                throw new \Exception($this->trans->t("Error occurred in the document service"));
-            }
-
-            $version = floatval($commandResponse->version);
-            if ($version > 0.0 && $version < 4.2) {
-                throw new \Exception($this->trans->t("Not supported version"));
-            }
-
-        } catch (\Exception $e) {
-            $this->logger->logException($e, ["message" => "CommandRequest on check error", "app" => $this->appName]);
-            return [$e->getMessage(), $version];
-        }
-
-        $convertedFileUri = null;
-        try {
-
-            $hashUrl = $this->crypt->GetHash(["action" => "empty"]);
-            $fileUrl = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".callback.emptyfile", ["doc" => $hashUrl]);
-            if (!empty($this->config->GetStorageUrl())) {
-                $fileUrl = str_replace($this->urlGenerator->getAbsoluteURL("/"), $this->config->GetStorageUrl(), $fileUrl);
-            }
-
-            $convertedFileUri = $documentService->GetConvertedUri($fileUrl, "docx", "docx", "check_" . rand());
-
-        } catch (\Exception $e) {
-            $this->logger->logException($e, ["message" => "GetConvertedUri on check error", "app" => $this->appName]);
-            return [$e->getMessage(), $version];
-        }
-
-        try {
-            $documentService->Request($convertedFileUri);
-        } catch (\Exception $e) {
-            $this->logger->logException($e, ["message" => "Request converted file on check error", "app" => $this->appName]);
-            return [$e->getMessage(), $version];
-        }
-
-        return ["", $version];
     }
 }