+++ l10n/bg.js
@@ -51,6 +51,7 @@ OC.L10N.register(
     "Select recipients" : "Изберете получатели",
     "Select tag" : "Изберете етикет",
     "version": "версия",
-    "Keep intermediate versions when editing": "Съхранявайте междинни версии при редактиране"
+    "Keep intermediate versions when editing": "Съхранявайте междинни версии при редактиране",
+    "Generate document preview": "Генериране на визуализация на документ"
 },
 "nplurals=3; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2);");