+++ l10n/ru.js
@@ -85,6 +85,7 @@ OC.L10N.register(
     "Original": "Исходный документ",
     "version": "версия",
     "Disable certificate verification (insecure)": "Отключить проверку сертификата (небезопасно)",
-    "Keep intermediate versions when editing": "Хранить промежуточные версии при редактировании"
+    "Keep intermediate versions when editing": "Хранить промежуточные версии при редактировании",
+    "Generate document preview": "Генерировать превью документов"
 },
 "nplurals=3; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2);");